refactor(syscalls): move alloc syscalls into new module
